feat(user-ui): refactoring discussion
