+++ packages/user-ui/src/pages/projects/_id/components/discussion/i18n.rs
@@ -1,0 +1,50 @@
+use bdk::prelude::*;
+
+translate! {
+    DiscussionTranslate;
+
+    title: {
+        ko: "토론",
+        en: "Discussion",
+    },
+
+    sub_title: {
+        ko: "주요 내용",
+        en: "Highlights"
+    }
+
+    video_discussion: {
+        ko: "화상 토론",
+        en: "Video Discussion"
+    }
+
+    time_table: {
+        ko: "일정표",
+        en: "Time table"
+    }
+
+    time: {
+        ko: "시간",
+        en: "Time"
+    }
+
+    activity: {
+        ko: "활동",
+        en: "Activity"
+    }
+
+    content: {
+        ko: "내용",
+        en: "Content"
+    }
+
+    participant: {
+        ko: "참여자",
+        en: "Participant"
+    }
+
+    involved: {
+        ko: "참여하기",
+        en: "Involved"
+    }
+}
