+++ packages/user-ui/src/pages/projects/_id/components/discussion/mod.rs
@@ -1,0 +1,6 @@
+mod controller;
+mod i18n;
+
+mod discussion;
+
+pub use discussion::DiscussionTab;
